Replace strings with byte slices and utf8.DecodeRune
When we read the dictionary into memory with ioutil, it returns a
[]byte. Previously, we were converting that to a string, and doing
everything with strings. Including, in the sort method, allocating a new
string with strings.ToLower, splitting that and sorting it, and joining
the result to return a new string.

With these changes, we now deal with []byte everywhere. The public
interface still uses strings, because that seems like the more natural
approach. Since the Open function will usually take care of reading the
dictionary and adding the words, we will usually be calling the internal
add() method, which will avoid unnecessary conversion between strings
and []bytes.

When iterating through the characters in the words, search text and
filter, we now use the utf8.DecodeRune() function. This will allow us to
continue supporting unicode characters; if we were to just iterate
through the bytes in the slices directly, that would not be
unicode-friendly. Doing this requires a bit of fiddliness with manually
adding the width of each rune to the current index, but it's fairly
straightforward and worth it.

I've also added some functions to support calling sort.Sort() on a
[]byte by making the latter an interface{}.