Remove the ToLower case folding step
While we currently deal with strings in a Unicode-clean way, folding
everything to lowercase may be the wrong thing to do for languages other
than English.

Scrabble has tiles for specific characters in specific forms, and we
should not be modifying them. Instead, the dictionary we use needs to be
made for use with Scrabble; it will contain only the characters usable
and valid in the game. If we do that, we will have no need for changing
the case.