Change test data expected results for not lowercasing everything
Because the parent commit changes how input data is treated, the
expected anagram counts we get from the test data are different. So this
commit updates those counts to what should be expected.

To determine these counts I wrote a separate slower anagram solver and
discovered that, due to a bug in the filtering code, we didn't even have
the correct counts before; results that included a filter letter that
was capitalized in the word were excluded.

These should be the correct numbers. Ultimately, it might be better to
have synthetic test data that includes human-verified counts and is
constructed to highlight potential problems. I think the more practical
solution is to just have separate code (like the tool I mentioned above)
that generates the correct results in a slower but more straightforward
way.