Replace strings with byte slices everywhere (#16)
* Replace strings with byte slices and utf8.DecodeRune

When we read the dictionary into memory with ioutil, it returns a
[]byte. Previously, we were converting that to a string, and doing
everything with strings. Including, in the sort method, allocating a new
string with strings.ToLower, splitting that and sorting it, and joining
the result to return a new string.

With these changes, we now deal with []byte everywhere. The public
interface still uses strings, because that seems like the more natural
approach. Since the Open function will usually take care of reading the
dictionary and adding the words, we will usually be calling the internal
add() method, which will avoid unnecessary conversion between strings
and []bytes.

When iterating through the characters in the words, search text and
filter, we now use the utf8.DecodeRune() function. This will allow us to
continue supporting unicode characters; if we were to just iterate
through the bytes in the slices directly, that would not be
unicode-friendly. Doing this requires a bit of fiddliness with manually
adding the width of each rune to the current index, but it's fairly
straightforward and worth it.

I've also added some functions to support calling sort.Sort() on a
[]byte by making the latter an interface{}.

* Move word sorting out of add()

* Remove the ToLower case folding step

While we currently deal with strings in a Unicode-clean way, folding
everything to lowercase may be the wrong thing to do for languages other
than English.

Scrabble has tiles for specific characters in specific forms, and we
should not be modifying them. Instead, the dictionary we use needs to be
made for use with Scrabble; it will contain only the characters usable
and valid in the game. If we do that, we will have no need for changing
the case.

* Change test data expected results for not lowercasing everything

Because the parent commit changes how input data is treated, the
expected anagram counts we get from the test data are different. So this
commit updates those counts to what should be expected.

To determine these counts I wrote a separate slower anagram solver and
discovered that, due to a bug in the filtering code, we didn't even have
the correct counts before; results that included a filter letter that
was capitalized in the word were excluded.

These should be the correct numbers. Ultimately, it might be better to
have synthetic test data that includes human-verified counts and is
constructed to highlight potential problems. I think the more practical
solution is to just have separate code (like the tool I mentioned above)
that generates the correct results in a slower but more straightforward
way.

* Add a benchmark for calling the Open() func