Move implementation into internal dir
This avoids having to support the API publicly. (The 'internal' dir is a
Go tooling mechanism, not an arbitrary name.) It also cleans up the repo
root a bit.

See here for more reasoning: https://eli.thegreenplace.net/2019/simple-go-project-layout-with-modules/

It isn't great to have the internal package named the same as the module,
but I'll leave it like this for now.