Remove existing tests and benchmarks
These tests are not unit tests, because the code is not separated well enough
to effectively test them as such. So they're de facto integration tests. But
they aren't good ones, because the test results are unverified. Better to
remove these than have false confidence. We can re-add tests that are better
once we've more effectively separated the code, and come up with better ways
to verify test results.